+++ src/main/java/com/gihae/shop/cart/repository/Cart.java
@@ -22,7 +22,7 @@ public class Cart {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
@@ -37,7 +37,7 @@ public class Cart {
     private int price;
 
     @Builder
-    public Cart(int id, User user, Option option, int quantity, int price) {
+    public Cart(Long id, User user, Option option, int quantity, int price) {
         this.id = id;
         this.user = user;
         this.option = option;
